python_ocr/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
pillow==10.0.1
pytesseract==0.3.10
openai==1.3.5
numpy==1.25.2
python-dotenv==1.0.0 